Move iwear utils in a common place
